refactor build collection component
