+++ constants/CollectionConstants.ts
@@ -1,2 +1,9 @@
 export const MINIMUM_WIDTH = 1200;
 export const MINIMUM_HEIGHT = 1200;
+
+export interface BuildCollectionInputType {
+  name: string;
+  slug: string;
+  description: string;
+  keywords: string;
+}
